[Refactoring ⚙️] 프로젝트 목록 분리
